Fix build break

+++ OpenSim/Region/CoreModules/Avatar/Gestures/GesturesModule.cs
@@ -85,7 +85,7 @@ namespace OpenSim.Region.CoreModules.Avatar.Gestures
             item = invService.GetItem(item);
             if (item != null)
             {
-                item.Flags &= ~1;
+                item.Flags &= ~(uint)1;
                 invService.UpdateItem(item);
             }
             else
